requirements.cd.txt
django == 3.2.3
django-environ == 0.4.5
django-filter
djangorestframework == 3.12.4
psycopg2-binary == 2.8.6
markdown
